Enforce full-semver pins via a gated CI check
Replace the always-on vitest spec with a dedicated CI check, gated to PRs
that change package.json. scripts/check-package-pins.mjs walks every registry
dependency and fails (naming each section / dependency / range) if a pin is
not a full major.minor.patch version; non-registry specifiers (URLs,
workspace:/catalog:/link:/file:, dist-tags) are exempt. Wired as the
"check:pins" pnpm script and run by the Package Pins workflow, which is
path-gated to package.json (plus the script and workflow), so it only runs on
PRs that touch packages. Register its job in the timeout enforcement test and
drop src/package-version-pins.spec.ts, whose role this check now fills.

Modeled on rmartz/ai-tools#63.
